Update app configs
