--- lib/site.ts
@@ -1,5 +1,0 @@
-export const site = {
-  title: "Nook++",
-  description: "My ACNH Wiki",
-  themeColor: "#82d7aa",
-};
